slimmed down the client wrapper to just include just what we need
