+++ src/test/testHooks.ts
@@ -1,5 +1,5 @@
 import { Context } from 'mocha';
-import { AppinsightsKey, JVSC_EXTENSION_ID, Telemetry } from '../platform/common/constants';
+import { JVSC_EXTENSION_ID, Telemetry } from '../platform/common/constants';
 import { IS_CI_SERVER } from './ciConstants.node';
 import { extensions } from 'vscode';
 import { sleep } from './core';
@@ -8,13 +8,14 @@ import { CiTelemetryReporter } from './utils/ciTelemetry/ciTelemetryReporter.nod
 let telemetryReporter: CiTelemetryReporter | undefined;
 
 export const rootHooks = {
-    beforeAll() {
+    beforeAll: async function () {
         if (!IS_CI_SERVER) {
             return;
         }
 
         const extensionVersion = extensions.getExtension(JVSC_EXTENSION_ID)?.packageJSON.version;
-        telemetryReporter = new CiTelemetryReporter(JVSC_EXTENSION_ID, extensionVersion, AppinsightsKey, true);
+        telemetryReporter = new CiTelemetryReporter(JVSC_EXTENSION_ID, extensionVersion);
+        await telemetryReporter.initialize();
     },
     afterEach(this: Context) {
         if (!IS_CI_SERVER) {
@@ -25,7 +26,7 @@ export const rootHooks = {
         if (this.currentTest?.title) {
             const duration = this.currentTest?.duration;
             const measures = typeof duration === 'number' ? { duration: duration } : duration ? duration : undefined;
-            telemetryReporter?.sendRawTelemetryEvent(
+            telemetryReporter?.sendTelemetryEvent(
                 Telemetry.RunTest,
                 {
                     testName: this.currentTest?.title,
@@ -40,7 +41,7 @@ export const rootHooks = {
             return;
         }
 
-        await telemetryReporter?.dispose();
+        await telemetryReporter?.flush();
         // allow some time for the telemetry to flush
         await sleep(2000);
     }
